+++ peepholelib/datasets/dataset_base.py
@@ -17,24 +17,6 @@ class DatasetBase(metaclass=abc.ABCMeta):
     def load_data(self):
         raise NotImplementedError()
     
-    def get_train_dataset(self):
-        if not self._train_ds:
-            raise RuntimeError('Data not loaded. Please run model.load_data() first.')
-
-        return self._train_ds
-    
-    def get_val_dataset(self):
-        if not self._train_ds:
-            raise RuntimeError('Data not loaded. Please run model.load_data() first.')
-
-        return self._val_ds
-                                 
-    def get_test_dataset(self):
-        if not self._train_ds:
-            raise RuntimeError('Data not loaded. Please run model.load_data() first.')
-
-        return self._test_ds
-    
     def get_classes(self):
         if not self._classes:
             raise RuntimeError('Data not loaded. Please run model.load_data() first.')