+++ RELEASENOTES.md
@@ -9,3 +9,4 @@
 - `ClassifierBase` can be saved and loaded. It receives `path` and `name` arguments for saving.
 - `Peepholes` objects now compute peepholes using a `DrillBase` class
 - `Peepholes` implements [Deep Mahalanobis Distance](https://arxiv.org/abs/1807.03888) as peephole computation
+- Framework now accepts any `torch.nn.Module` for computing `corevectors` and `peepholes`, appropriate functions need to be provided. `target_layers` now are `target_modules`